+++ piargus/__init__.py
@@ -4,18 +4,16 @@ from piargus.batchwriter import BatchWriter
 from piargus.codelist import CodeList
 from piargus.constants import *
 from piargus.treerecode import TreeRecode
-from piargus.hierarchy import Hierarchy
-from piargus.treehierarchy import TreeHierarchy, TreeHierarchyNode
-from piargus.codehierarchy import CodeHierarchy
-from piargus.inputdata import InputData
+from piargus.hierarchy import Hierarchy , TreeHierarchy, TreeHierarchyNode , CodeHierarchy
+from piargus.inputdata import InputData, MetaData, MicroData, TableData
 from piargus.job import Job
 from piargus.job import JobSetupError
-from piargus.metadata import MetaData
-from piargus.microdata import MicroData
+# from piargus.metadata import MetaData
+# from piargus.microdata import MicroData
 from piargus.safetyrule import dominance_rule, percent_rule, frequency_rule, request_rule, zero_rule, \
     missing_rule, weight_rule, manual_rule, p_rule, nk_rule
 from piargus.table import Table
-from piargus.tabledata import TableData
+# from piargus.tabledata import TableData
 from piargus.tableset import TableSet
 from piargus.tauargus import TauArgus