+++ piargus/hierarchy/__init__.py
@@ -1,0 +1,7 @@
+from .codehierarchy import CodeHierarchy
+from .flathierarchy import FlatHierarchy
+from .treehierarchy import TreeHierarchy, TreeHierarchyNode
+from .hierarchy import Hierarchy
+
+__all__ = ["Hierarchy", "TreeHierarchy", "CodeHierarchy", "FlatHierarchy", "TreeHierarchyNode"]
+