+++ piargus/inputdata/__init__.py
@@ -1,0 +1,6 @@
+from .inputdata import InputData
+from .microdata import MicroData
+from .tabledata import TableData
+from .metadata import MetaData
+
+__all__ = ["InputData", "MicroData", "TableData", "MetaData"]