+++ piargus/tableset.py
@@ -1,3 +1,5 @@
+# TODO Remove
+
 import warnings
 from typing import Mapping, Hashable, Union, Iterable, Optional, Sequence